+++ Assets/Scripts/Shooting, Laser, & Damage/OldSpawnLaserOnInput.cs
@@ -16,7 +16,7 @@ public class OldSpawnLaserOnInput : MonoBehaviour
 
     [SerializeField] private float _shotCooldownDuration = .5f;
     [SerializeField] private float _laserPushForce = 5;
-    [SerializeField] private float _laserDamage = 1;
+    [SerializeField] private int _laserDamage = 1;
 
     [SerializeField] private FlashBarrelOnFire _barrelFlashScriptRef;
